Added .gitignore, separated prompt from class
